Fix cub-less device info compilation
